# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 为API生，为框架死，为debug奋斗一辈子；吃符号亏，上大小写的当，最后死在需求上   ------- 梦旋飞
  version: 1.0.0
  title: 开发平台
basePath: /channel_interface
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /taobao/getGoods:
    get:
      tags:
        - 公共接口
      summary: 调用淘宝接口获取商品信息(https://detail.tmall.com/item.htm?id=569505623525)
      produces:
        - application/json
      parameters:
        - name: url
          description: 淘宝url
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /taobao/getShop:
    get:
      tags:
        - 公共接口
      summary: 调用淘宝接口获取店铺信息（https://fengpailu.tmall.com/shop/view_shop.htm?user_number_id=3828845976）
      produces:
        - application/json
      parameters:
        - name: url
          description: 店铺url
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'







  #数据接口
  #首页运营数据接口
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...